+++ src/config.rs
@@ -152,6 +152,7 @@ pub fn get_config_args(
         wrapmin: config
             .get("wrapmin")
             .map(|x| x.as_integer().unwrap().try_into().unwrap()),
+        join: config.get("join").map(|x| x.as_bool().unwrap()),
         tabsize: config
             .get("tabsize")
             .map(|x| x.as_integer().unwrap().try_into().unwrap()),